requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
werkzeug
argon2-cffi
python-dotenv
